refactor: split app and command into their own modules
