+++ packages/go/trynoports/app/app_view.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"fmt"
