+++ packages/go/trynoports/command/go.mod
@@ -1,0 +1,3 @@
+module command
+
+go 1.22.2
