+++ packages/go/trynoports/go.work
@@ -1,0 +1,7 @@
+go 1.22.2
+
+use (
+	.
+	./app
+	./command
+)
